feat(chart): add chart config context variable
