+++ frontend/src/app/components/FormGenerator/Layout/CollectionLayout.tsx
@@ -41,34 +41,46 @@ import { groupLayoutComparer } from '../utils';
 import ItemLayout from './ItemLayout';
 
 const CollectionLayout: FC<FormGeneratorLayoutProps<ChartStyleSectionConfig>> =
-  memo(({ ancestors, translate, data, dataConfigs, flatten, onChange }) => {
-    const getDependencyValue = useCallback((watcher, children) => {
-      if (watcher?.deps) {
-        // Note: only support depend on one property for now.
-        const dependencyKey = watcher?.deps?.[0];
-        return children?.find(r => r.key === dependencyKey)?.value;
-      }
-    }, []);
+  memo(
+    ({
+      ancestors,
+      translate,
+      data,
+      dataConfigs,
+      flatten,
+      onChange,
+      context,
+    }) => {
+      const getDependencyValue = useCallback((watcher, children) => {
+        if (watcher?.deps) {
+          // Note: only support depend on one property for now.
+          const dependencyKey = watcher?.deps?.[0];
+          return children?.find(r => r.key === dependencyKey)?.value;
+        }
+      }, []);
 
-    return (
-      <StyledCollectionLayout>
-        {data?.rows
-          ?.filter(r => Boolean(!r.hide))
-          .map((r, index) => (
-            <ItemLayout
-              ancestors={ancestors.concat([index])}
-              key={r.key}
-              data={r}
-              translate={translate}
-              dependency={getDependencyValue(r.watcher, data?.rows)}
-              dataConfigs={dataConfigs}
-              flatten={flatten}
-              onChange={onChange}
-            />
-          ))}
-      </StyledCollectionLayout>
-    );
-  }, groupLayoutComparer);
+      return (
+        <StyledCollectionLayout>
+          {data?.rows
+            ?.filter(r => Boolean(!r.hide))
+            .map((r, index) => (
+              <ItemLayout
+                ancestors={ancestors.concat([index])}
+                key={r.key}
+                data={r}
+                translate={translate}
+                dependency={getDependencyValue(r.watcher, data?.rows)}
+                dataConfigs={dataConfigs}
+                flatten={flatten}
+                onChange={onChange}
+                context={context}
+              />
+            ))}
+        </StyledCollectionLayout>
+      );
+    },
+    groupLayoutComparer,
+  );
 
 export default CollectionLayout;
